wifi: Report timestamp differences in syslog messages
When testing auto-reassociation (in this case due to deauth)
messages, most of the work happens behind the scenes and is
not logged by supplicant/flimflam.  In order to keep metrics
on these events, we can pull out events from the logs.  Do
this for suspend->deauth->reassociate->COMPLETED in all the
SuspendTimeout tests.  This will add the following perf
results to successfully run tests:

  log_time_diff_00_deauth       0.603645000001
  log_time_diff_01_reassoc      0.607721999999
  log_time_diff_02_reconnect    0.00317597389221

This version does not handle log events that span mutliple
rotated files, but it works well enough in the simple case
to be useful for collecting data samples.

BUG=chromium-os:14037
TEST=Rerun Roaming suite (also some manual tests related to timestamp
parsing)

Change-Id: Iae93eef21e98e99ff573cb7e22ba16908db77a5d
Reviewed-on: http://gerrit.chromium.org/gerrit/850
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>