hardware_StorageFio: only use rootdev if we are really testing boot disk
When testing hardware_StorageFio, the device may look like a root device,
especially on ARM device (mmcblk*/sda*). hardware_StorageFio used to trust a
fixed list to decide if we're testing with internal storage; however that would
be wrong on ARM platforms.

This CL checks the boot device in a more currect way - the "rootdev -s -d"
command.

BUG=chrome-os-partner:3882
TEST=Performed test successfully.

Change-Id: Ib0eaba351aadbc578aacdb78d65d0cd6e1121cc3
Reviewed-on: http://gerrit.chromium.org/gerrit/1096
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>