Add security_RootCA test for validating roots shipped in CrOS
BUG=chromium-os:9017
TEST=run_remote_tests ... this test currently passes on 0.13.525
The test currently passes, unsurprisingly, since the current baseline
is derived from current shipping reality. This has been previously manually
inspected vs the "golden" list and approved, but it wouldn't
hurt to do that one more time coinciding with this check in, since it's been
a few months. I'll take care of that and log any discrepancies with
bugs.

Change-Id: I0cb0edbd9e8911c47145fc9e36bacf340005b54b
Reviewed-on: http://gerrit.chromium.org/gerrit/1122
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Tested-by: Jim Hebert <jimhebert@chromium.org>