autotest: power_Resume: fix test for ARM
On ARM systems, the dmesg timestamp jumps on resume by the time that the
system has been suspended for. On x86 systems, this is not the
case. Due to this, the resume time reported on ARM systems by
power_Resume test is incorrectly inflated by the amount of time the
system has been suspended for.

Fixed the test to ignore time spent in suspended state.

BUG=chrome-os-partner:3250
TEST=run this test on x86 and ARM systems and ensure that the suspend and
resume times stored in keyval files are reasonable.

Change-Id: I04e412216a026db1fdd7f7cd83a69c11847451c3
Reviewed-on: http://gerrit.chromium.org/gerrit/1135
Tested-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>