Rewrite log_time_diff to use downloaded script
Take an earlier suggestion to both (a) use a downloaded script instead
of a lengthy ssh interaction (b) handle bits like wrapping around to
the last log-rotated file and automatically rejecting any log entries
that fall before the start of the current test.

BUG=chromium-os:14037
TEST=Run 004* roaming tests a bunch of times.  There are still remaining
issues with the test itself (DUT doesn't get reliably deauthed) but as
for the logfile parsing itself, it is looking better.

Change-Id: I2c7d40d66fcb74421861187687e4fc381e64ad08
Reviewed-on: http://gerrit.chromium.org/gerrit/1378
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>