Rewrite SuspendTimeout to do explicit deauth
The previous version of the SuspendTimeout test relied on setting
an inactivity timeout on the AP.  This turns out to be somewhat
unreliable as the AP does some number of attempts to restore
connectivity.  Instead, use the hostapd_cli interface to explicitly
deauth the DUT while it sleeps.

BUG=chromium-os:15689
TEST=Rerun -- all SuspendTimeout tests and ensure that not only do tests
pass (pings succeed) but that all the log_time_diff values appear

Change-Id: Ibb88e34a2d8150ace74cca2b0df3cf01bea8808c
Reviewed-on: http://gerrit.chromium.org/gerrit/1468
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>