[autotest] Force FQDN when trying to DNS-resolve www.google.com in CrosUITest setup
In the setup phase of CrosUITest, we fake out DNS by re-routing all
DNS resolution requests to a local DNS server.  We check that this has
been done by attempting a DNS resolution of 'www.google.com' and
expecting to get back 127.0.0.1.  By forcing the use of a
fully-qualified-domain-name, we make the resolution attempt go
directly to the DNS server, bypassing any local potential sources of
failure or hanging.

BUG=chromium-os:13384
TEST=suite_Smoke

Change-Id: I58f0f7175db6f0e2001fe164c5935ef7e5a18e4d
Reviewed-on: http://gerrit.chromium.org/gerrit/1495
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>