Wait for crash notification message to appear in logs.
There appears to a be a race condition in the logging system. Messages
sent to syslog may take a moment before appearing in the written log.

Due to the way it was written logging_UserCrash was expecting pretty
tight timings here. I've modified the code to allow for a 5 sec window
which should prevent any flake.

BUG=none
TEST=Many iterations of logging_UserCrash.

Change-Id: I834c35c992696748b4fe0c6bc51ce2146597bcdf
Reviewed-on: http://gerrit.chromium.org/gerrit/1514
Tested-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>