[autotest] Switch to getaddrinfo and print route table on DNS checks
gethostbyname is deprecated.  Switch to using getaddrinfo in the hopes of
getting better behavior.  Also, log the device routing table on every
DNS resolve request that occurs while we're trying to switch to/from
the local, fake, DNS server.  This info will hopefully help diagnose
the failures that will, in all likelihood, continue to occur even after
the switch to getaddrinfo.

BUG=chromium-os:13384
TEST=suite_Smoke

Change-Id: I07d1dc5a4e0292d6f0fecc5f3627db63bef9355e
Reviewed-on: http://gerrit.chromium.org/gerrit/1896
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>