Remove unnecessary sync calls from setup phase.
These three tests run 'sync' during the setup phase of the build. This causes
the build to pause and wait for all pending writes to complete. This step
delays the build unnecessarily. In my tests I saw this step causing the build
to hang for several minutes because other processes were writing to the disk.

I ran a build both before and after this change. Before the change took
14m7.6s and after it took 6m3.154s (because it wasn't waiting around).

BUG=chromium-os:16218
TEST=Verify autotest-tests still works and compiles faster.

Change-Id: I6b9e0a82033e71414a88995986878eb5edd5c641
Reviewed-on: http://gerrit.chromium.org/gerrit/2188
Reviewed-by: Eric Li <ericli@chromium.org>
Tested-by: David James <davidjames@chromium.org>