wifitest: Add new "StrictMode" test
Test to make sure that a client can successfully re-authenticate to
a strict-auth AP that it has lost synchronization with.  I made up
the term "strict auth" to refer to APs that take the IEEE guidance
to heart that if a client sends an AUTH packet and there is state
on the AP to indicate the client is already authenticated, that it
must send a DEAUTH downstream first indicating that the old
authentication is now no longer valid.  We set up an AP that
replicates this behavior, and duplicate the scenario by rebooting
the client while it is associated.

BUG=chromium-os:14765
TEST=Run test

Change-Id: Ide71409b682e23a7634823701d10b9e558e9ae18
Reviewed-on: http://gerrit.chromium.org/gerrit/2225
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>