Move host randomization and ssh checks into new metahost scheduler.
The previous implementation wasn't actually randomizing the host list
since the set conversion was wiping out the shuffle. Creating a new
metahost scheduler is the most practical approach to fixing this.

A new metahost scheduler has the added benefit of being more
understandable as well. The downside is a good portion of the code is a
copy paste from the existing LabelMetahostScheduler.

BUG=none
TEST=Running in production.

Change-Id: Ic26dff1ee287b8c2656680ad900cd1da3e0df39a
Reviewed-on: http://gerrit.chromium.org/gerrit/2364
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Dale Curtis <dalecurtis@chromium.org>