wifi: Automatically determine signal threshold
The SSIDMultiSwitch test has been plagued with an arbitrary
signal strength threshold which is insensitive to the signal
conditions of the individual test cell / device reporting.

This CL removes that constant and instead samples current
conditions in order to devise a threshold that will presumably
trigger when we drop the AP signal strength.

BUG=chromium-os:16342
TEST=Rerun test with or without antennae in one side of the test
cell (40 db difference)

Change-Id: I77a122640afa37920ae266f367f72d130fef05fb
Reviewed-on: http://gerrit.chromium.org/gerrit/2380
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>