Fix host=client test parameter errors.
logging_KernelCrashServer is failing with the error:
    AttributeError: 'str' object has no attribute 'wait_up'

This come from the test passing a {'host': 'client'} parameter into
distribute_across_machines.  The client parameter is unnecessary since
autotest defaults to using the client host already.  So I've removed
this from the control file, and added a sanity check in
server_job_utils.test_item.__init__(...) to remove this parameter if
it exists in case it is accidentally added back in any other tests in
the future.

BUG=None.
TEST=Local test runs.

Change-Id: I51327e3a0f4f65a1cd248b86dd261c5b0109d9d9
Reviewed-on: http://gerrit.chromium.org/gerrit/2447
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Paul Pendlebury <pauldean@chromium.org>