Added cros_perf profiler which passes command-line options directly to "perf record".
In the perf profiler, the programmer specified options to perf through python arguments. These
options were then converted to a perf command line. It is much more flexible and inuitive to allow the
programmer to specify the
perf command line options directly. This CL adds the cros_perf profiler which does that. This CL also adds a
check to see if perf is still running when the
autotest finishes. If perf isn't still running, something probably went wrong.

BUG=15811
TEST=Ran numerous autotests with the profiler and checked the options were being passed correctly. Also
checked that if invalid options were passed to perf, the test would fail.

Change-Id: I50aae6e141d130edff1adf3eee984e24f5e565dd
Reviewed-on: http://gerrit.chromium.org/gerrit/2526
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>