PlaybackRecord: fix parsing of 'aplay -l' on ZGB.
Set a default value for the 'direction' of a given control. This
allows the test to handle output of 'amixer' that includes controls
that are neither Playback nor Capture without giving a key error when
looking it up.

For example having this in the output of amixer would have caused the
test to stop.
Simple mixer control 'Headphone',0
  Capabilities: pswitch penum
  Playback channels: Front Left - Front Right
  Mono:
  Front Left: Playback [on]
  Front Right: Playback [on]

BUG=none.
TEST=Can now run this test on a ZGB.

Change-Id: I41f8eac42c4f196552fa37b463d0593dbabedc67
Reviewed-on: http://gerrit.chromium.org/gerrit/3383
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>