Add server_job execution ability to distribute jobs.
(Patch 4: Fix line over 80 chars.)

(Patch 3: Make server_job an optional argument so the gtest_runner
    tests will not need to be modified to continue using test_item.)

Some tests need to run as server jobs instead of client jobs.  Adding
to the existing test attributes that specify a test needs to reboot
before or after running to specify that a test needs to run as a
server job.  This fixes the logging_KernelCrashServer failures.

BUG=None.
TEST=Local: http://pauldean.kir/afe/#tab_id=view_job&object_id=323

Change-Id: I846f995417e9a39d2bd193a54689f9945772c033
Reviewed-on: http://gerrit.chromium.org/gerrit/2836
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Paul Pendlebury <pauldean@chromium.org>