Audio loop factory test
The utility called 'audioloop' allows audio to loop from a chosen input device to a chosen output device until a sigterm is caught.

This utility (packaged in the form of an autotest dependency) is used by the factory_Audio factory test to perform a quality check on hardware by allowing a signal that is fed in through an external signal generator to be looped out through headphones and compared to the original signal.

BUG=None
TEST=See above
Change-Id: I96d49f07d2f0df10c419089cc610f44fc1716e85
Reviewed-on: http://gerrit.chromium.org/gerrit/3830
Reviewed-by: Taylor Hutt <thutt@chromium.org>
Tested-by: Abhishek Shroff <shroff@chromium.org>