Collect update engine logs on failure for debugging.
There are many instances where we need to check the update_engine logs
after a job has failed, but generally the machine has already been
reimaged by the time we notice a failure. Since the error occurs in the
install phase logs are never collected.

Instead of copying all logs, I've just made it so the update engine logs
and the latest messages file are retrieved. I chose to do it this way
since the system is in an unknown state prior to imaging, there could be
a large quantity of logs and crashes we aren't interested in. We can
re-evaluate this decision going foward if there's missing information.

BUG=none
TEST=Ran updater, killed update-engine, checked logs.

Change-Id: I1dba7d6016f4020c6173ae22db3a74af038d964a
Reviewed-on: http://gerrit.chromium.org/gerrit/3994
Tested-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>