network_3GSuspendResume: test with autoconnect on and off
network_3GSuspendResume: test with autoconnect on and off.  The this
test was flaky because if autoconnect was on the test would fail, but
if it were off, the test would pass.  Make this test more meaningful
by ensuring that we know the state of the autoconnect property and
running all the scenarios both ways.

BUG=chromium-os:17556
TEST=./run_remote_tests.sh --build network_3GSuspendResume --remote=

Change-Id: I13b128d80d99af332ca5cc2a941df7ab09e24489
Reviewed-on: http://gerrit.chromium.org/gerrit/4029
Reviewed-by: ttuttle <ttuttle@chromium.org>
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>