network_3GSuspendResume: test that autoconnect works after suspend
network_3GSuspendResume: test with autoconnect works after suspend, by
checking for several suspend resume cycles that we autoconnect when
autoconneect is True, and that we do not autoconnect if it is False.
Change the suspend behavior of the test so that it uses the path that
is 'normally' used when suspend.  This path will causes cromo to
disconnect and existing data sessions.

BUG=chromium-os:17556
TEST=./run_remote_tests.sh --build network_3GSuspendResume --remote=

Change-Id: Iecd8558b2bd447f87b02683a6a70ac1d70ba6f3e
Reviewed-on: http://gerrit.chromium.org/gerrit/4045
Tested-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>