[autotest] Address failure to cache DNS settings in CrosUITest autotests
We used to attempt to cache live DNS settings at the beginning of
all autotests that inherit from CrosUITest, and then re-set them
at the end of the test.  This can be prone to races with flimflam's
discovery of DNS settings.  Instead, we'll just restart flimflam at
the end of the test, and allow it to re-discover DNS settings.

BUG=chromium-os:17427
TEST=suite_Smoke

Change-Id: Ie2d6b897af5d9721982ddd17515d6d3f750e6113
Reviewed-on: http://gerrit.chromium.org/gerrit/4260
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>