hardware_Trackpad: use an override-redirect popup window for Xcapture module
Currently, xev is used to capture X events. When xev was launched, its
geometry was controlled by the window manager, and did not cover the
whole screen. In this CL, an override-redirect popup window is created
so that its geometry can be easily manipulated to cover the whole screen.
And then xev can listen to the popup window without missing X events.

This CL had been reverted and is now re-submitted with using the correct
method of getting an autox instance.

BUG=chromium-os:14060
TEST=From host, install hardware_Trackpad in the chromebook, and check
that xev covers the whole screen.
$ ./run_remote_tests.sh --remote={CHROMEBOOK_IP} hardware_Trackpad/control

Change-Id: I80c3bf34671230ddef8c52b3b45ba424492edc23
Reviewed-on: http://gerrit.chromium.org/gerrit/4993
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>