hardware_Trackpad: use an override-redirect popup window for Xcapture module
Currently, xev is used to capture X events. When xev was launched, its
geometry was controlled by the window manager, and did not cover the
whole screen. In this CL, an override-redirect popup window is created
so that its geometry can be easily manipulated to cover the whole screen.
And then xev can listen to the popup window without missing X events.

BUG=chromium-os:14060
TEST=From host, install hardware_Trackpad in the chromebook, and check
that xev covers the whole screen.
$ ./run_remote_tests.sh --remote={CHROMEBOOK_IP} hardware_Trackpad/control

Change-Id: I7753a18acea9aef637c8440b5491d583306f260c
Reviewed-on: http://gerrit.chromium.org/gerrit/4565
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>