network_3GDormancyDance: Wait for service to appear before connecting
There was a race condition in 3GDormancyDance: it tried to connect
immediately after the modem's Powered property changed to True.  Most
of the time, this would work, but sometimes it would attempt to connect
to the cell service before it existed.  It would fail (silently),
leaving the modem disconnected but the test expecting it to connect.

This patch makes the test wait for the cell service to appear before
trying to connect to it.

BUG=chromium-os:15229
TEST=Adhoc; test now runs to completion (Quis custodiet ipsos custodes?)

Change-Id: I5a907234c50bc8ce22169f8c4f3d9407a4f7af69
Reviewed-on: http://gerrit.chromium.org/gerrit/4582
Reviewed-by: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>