network_3GDormancyDance: Handle when modem is initially disabled
If the cell service is not available when we set up the test, assume the
modem is disabled and start by enabling it.  This does not cover all cases,
but makes the test slightly more robust.  (Ideally, we should check for
all of the possible states, and just jump in to the test state machine at
whatever point the modem is.)

BUG=chromium-os:18062
TEST=Adhoc; test starts properly with modem initially disabled

Change-Id: Iae915d6207e2b6ea61c015e95431fe2609121f20
Reviewed-on: http://gerrit.chromium.org/gerrit/4604
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>