network_3GDormancyDance: Add a real state machine, finally.
DormancyDance has been in dire need of some actual logic for tracking
what the modem is doing, instead of just knee-jerk responding to each
incoming event.

I've added an enumeration of States, split out the handling of each
incoming event, and added exceptions for unexpected transitions.
The test still hangs on exceptions, because it's not using an
exception-forwarding main loop; I'll change it to use one in a future
patch.

BUG=chromium-os:17848
TEST=DormancyDance passed with ops=500 as a cursory smoketest

Change-Id: I50de82795a16ea6a47d5cb9635482abcb482502e
Reviewed-on: http://gerrit.chromium.org/gerrit/5488
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>