autotest: wifi: A new test for preferring 5 GHz channels.
This tests creates 2 APs, one with a 2.4 GHz channel and another with a 5 GHz
channel.  They are set to the same power level.  The DUT should consistently
connect to the 5 GHz channel.

Added a new check: client_check_frequency, which verifies that the DUT is using
the expected channel.

This test depends on a corresponding change in wpa_supplicant in Change
I804a6ac7.

BUG=chromium-os:18867
TEST=Ran this test with a build containing Change I804a6ac7 and it passes.
Fails without it.

Change-Id: I39dfd5a478dbaef328a3e698171c24d6ebb92ab2
Reviewed-on: http://gerrit.chromium.org/gerrit/5577
Reviewed-by: Gary Morain <gmorain@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>