network_3GDormancyDance: forward exceptions
Use ExceptionForwardingMainLoop to forward exceptions thrown by methods
in the test to the top level.  This makes the test terminate instead of
hanging when something breaks. :)

Also, count entire loops of the test instead of individual ops, and lower
it to 20 loops (~80 ops) instead of the 5000 that we used to have.  (We
might want to raise this for the autotests, but it is a good length for
manual testing.)

BUG=chromium-os:17848
TEST=Adhoc; network_3GDormancyDance passes and fails as expected

Change-Id: I070a90868af5e5df833c70c299ad569564a710ff
Reviewed-on: http://gerrit.chromium.org/gerrit/5794
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>