network_3GSafetyDance: Ensure Favorite=True before setting AutoConnect=False
If we run 3GSafetyDance on a newly-imaged machine, we've never
connected to Verizon, so Verizon is not set as a favorite service, so
flimflam won't let us set AutoConnect on it, so when we try to disable
AutoConnect, flimflam gives us an InvalidService error.

This patch adds an extra workaround that makes sure Favorite=True
before trying to set AutoConnect=False.  (It also asserts that the
final state of the service is what we expect.)

BUG=chromium-os:18981
TEST=Adhoc; reimaged a machine and ran SafetyDance, and it passed

Change-Id: I17b0bf796e7d486c921a53c81c62c8c2d9e28794
Reviewed-on: http://gerrit.chromium.org/gerrit/5679
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>