suite_Factory: recovery factory UI from corrupted state
The state control of factory is currently handle by state server, and no need to
rely on autotest's state file - except the "reboot" state.

This CL tries to add a new "reboot flag" file so that init scripts can know if
we should enter "continue" mode only if the system was scheduled for a reboot.

In init script side, reboot flag must be cleared for each start.

BUG=chrome-os-partner:3107
TEST=start factory ui, unplug power, reboot * 10; still see UI.

Change-Id: I31822aa3d715c27160ad2ff4f2a27d3ee0a0c0e8
Reviewed-on: http://gerrit.chromium.org/gerrit/5727
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>