Parse reason out of logs for display in results emails.
A recent upstream Autotest merge has reduced the effectiveness of
reporting error logs alone. With this patch I've added the contents of
the reason field to the beginning of the log data.

Due to this change, I've reduced the number of lines returned from the
error log from 25 to 10.

I also fixed a parser error which caused reasons with an equal sign in
them to be treated as optional parts.

BUG=none
TEST=ran parser against results. running in production.

Change-Id: I2a5970013f1b8b04f21860af66974a5fb65ba864
Reviewed-on: http://gerrit.chromium.org/gerrit/5905
Reviewed-by: Nirnimesh <nirnimesh@chromium.org>
Tested-by: Dale Curtis <dalecurtis@chromium.org>