hardware_Trackpad: support specifying wildcard (*) in sequence criteria
This CL supports a new syntax to add wildcard in the sequence criteria.
For simplicity to satisfy current requirement of detecting button wheel
events following any other arbitrary sequence of events, the wildcard can
be specified in the beginning of a sequence criteria. When wildcard is
used in the sequence criteria, it matches any number of detected X events.
Currently, the wildcard is only supported in the beginning of sequence.

BUG=chromium-os:19088
TEST=Run trackpad autotest program and observe the results of
the scroll-regardless_prior_state gesture files.

Change-Id: I3becdc2bedb4498baebdcae69ac090bfba7bdc12
Reviewed-on: http://gerrit.chromium.org/gerrit/6220
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>