autotest: Add site_cisco_router
Add the ability to configure a Cisco router with the same API as
the other site_*_router implementations.  This one is special in
that we need to ssh-tunnel through a Linux SSH "server" entity.
We also keep a shell open in IOS instead of running separate ssh
invocations for each command.

TODO: This is still a work in progress -- MatFunc mostly works,
except for a couple of esoteric tests, and SecMat also works, to
the extent that Cisco supports the various functions.  I'd like
to wait and do Roaming as a separate CL.

BUG=chromium-os:19327,chromium-os:19327
TEST=MatFunc, SecMat

Change-Id: Ia83bba77f33084ceec688a2d2426e5553e2a277e
Reviewed-on: http://gerrit.chromium.org/gerrit/6346
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>