hardware_Trackpad: decouple the test logic from functionality name
This CL decouples the test logic from the functionality name in order
that the test logic completely depends on the criteria specified in
the config file.

In this way, any new functionality can be added without the need to
modify the code if the functionality can be fully characterized by a
combination of existent criteria.

BUG=chromium-os:19334
TEST=Run trackpad autotest on a Chromebook. It should finish normally.

Change-Id: I856450476e7ef3f4131851193b3d40637b98760f
Reviewed-on: http://gerrit.chromium.org/gerrit/6221
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>