[autotest] Make ServiceLogin mock more full featured.
This patch provides a submit element for the fake login page served by the
authserver, so that we can inject some JS and actually submit the form
once the appropriate fields are populated.  The form also calls
gaia.chromeOSLogin.onAttemptedLogin onSubmit now, which I think is the
desired behavior.

Moreover, I have added a server-side handler that processes the submission
of the form, sending back a redirect to the continue URL provided when the
form is POSTed

BUG=chromium-os:18289
TEST=suite_Smoke with webui disabled, LoginSuccess with webui enabled -- see that a form with two empty text fields and a submit button shows up in the login pane.

Change-Id: Ic3f3d33434645229162cf42fdb279ab1221d135f
Reviewed-on: http://gerrit.chromium.org/gerrit/6648
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>