autotest: Create consent file, and chown it, atomically
Create consent file as a temp file first, chown it to chronos, and then
rename it to "Consent To Send Stats".  I'm not sure if this is necessary,
but it will fix the problem if there's somehow a race condition with
session_manager_setup.sh removing the consent file because it's owned by
root.

BUG=chromium-os:18413
TEST=logging_{UserCrash,CrashSender}

Change-Id: I805c2bac8526cbef94082e5ede1b1a87231773b2
Reviewed-on: http://gerrit.chromium.org/gerrit/6353
Reviewed-by: Ken Mixter <kmixter@chromium.org>
Tested-by: Michael Krebs <mkrebs@chromium.org>