Attempt to fix dying SSH connections.
Lots of jobs are showing up with the server connection dying yet the
test completing normally on the client. The issue could be related to
the flimflam restart we're issuing at the end of tests to clear the DNS
settings.

It's also possible there are some intermittent network issues
in the lab.

In either case extending the ServerAliveInterval should resolve the
problem.

BUG=none
TEST=running in production.

Change-Id: I0ba5142d93db8158230ea422faf915c42fc48ba6
Reviewed-on: http://gerrit.chromium.org/gerrit/6549
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Dale Curtis <dalecurtis@chromium.org>