[autotest] Disable flimflam captive portal checking during UI tests
In UI autotests, we tell flimflam to force the entire system to use a
locally running DNS server.  This causes flimflam to kick off a check
to see if the system is in a captive portal environment.  We actually
have a locally-running web server serving 204s at the appropriate URL,
but flimflam's http request times out for some reason.  The server
never sees the request.  Instead of digging into this, just disable
captive portal checking for all UI autotests.

If doing so becomes a problem, we can figure it out then and/or fix
the behavior in shill.

BUG=chromium-os:19344
TEST=suite_Smoke

Change-Id: If7669ed58604430a905d387823c04e1e552d6c1b
Reviewed-on: http://gerrit.chromium.org/gerrit/6628
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>