[autotest] Allow different delays of power button release in Servo
Different CrOS platforms have slightly different circuitry on their
power buttons.  These circuits create different delays for 'release'
which need to be accounted for in Servo's autotest power_key method.
This CL addresses the variability by including a sleep and re-read
loop.  A TODO has also been added for to address thoughts that it
might be more preferable to track these delays by platform in some
sort of data structure.

BUG=chrome-os-partner:5743
TEST=platform_LongPressPower run on platform which previously failed
and not it passes.

Change-Id: If51998c1da66426fda9c1cfb4a80743566fb5c43
Reviewed-on: http://gerrit.chromium.org/gerrit/6821
Reviewed-by: Nirnimesh <nirnimesh@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>