Fix django 1.3 DeprecationWarning for DATABASES.
Django 1.3 which we migrated to, modified the settings structure so that
the database settings are no longer a set of constants.  They are replaced
with a dictionary of dictionaries where the 'default' dictionary contains
the desired settings.  I've switched to that format to avoid the warning. Not
sure when those constants will actually stop working.

This is the actual message that no longer shows:
/usr/local/autotest/site-packages/django/db/__init__.py:19: DeprecationWarning:
settings.DATABASE_* is deprecated; use settings.DATABASES instead.

TEST=ran dashboard generation on local autotest

Change-Id: I647f2c353b67a0ec820895c7973bfffab84f3f87
Reviewed-on: http://gerrit.chromium.org/gerrit/6835
Reviewed-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>
Tested-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>