Add retrieving of crash dumps at the end of the test.
Currently, we only retrieve the crash dumps after a test iteration.
This does not catch crash dumps that are generated during the logout.
Retrieving the crash dumps at the end of the test will take care of
this.

BUG=chromium-os:18943
TEST=Manually force desktopui_ChromeFirstRender to fail while running
suite_Smoke and verify that crash dump is associated with correct test

Change-Id: I5247e686c0835a498377f0ddbd3969a2501cec69
Reviewed-on: http://gerrit.chromium.org/gerrit/6921
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>