Fix package checksum issue resulting in refetch every time.
When using run_remote_tests, tests are not prepackaged, which results in
a tests and deps being bundled and fetched by the client each time it
needs them. These packages are cached locally, but because the checksum
wasn't updated upon download they are re-requested each time.

Fixing this bug results in a 15% speedup in the Smoke suite, from 8m26s
to 7m21s.

BUG=none
TEST=Ran smoke.

Change-Id: I4832d4a5c8f80962672045dbc5da60ee23c21e1d
Reviewed-on: http://gerrit.chromium.org/gerrit/7047
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Dale Curtis <dalecurtis@chromium.org>