hardware_Trackpad: more flexible algorithm for sequence criteria
The old algorithm could only handle the case where the detected
x event sequence is precisely the same as the sequence criteria.
There were some shortcomings with the above restriction, when
some X events are optional or tolerable depending on how the
criteria are specified. Such events include Motion event and Button
Wheel events. Problems arise with the following two situations.

1. There are detected Motion/Button Wheel events, but they are
   not explicitly specified in sequence criteria. Instead, global
   criteria could have been set for all functionalities to address
   how much Motion and Button Wheel events could be tolerated.

2. There are no detected Motion/Button Wheel events, but they are
   explicitly specified in sequence criteria.

This CL fixes the criteria matching algorithm to accommodate the
two conditions above.

BUG=chromium-os:19088
TEST=Run trackpad autotest program. It should execute normally
on palm-* and scroll-* and other gesture files.

Change-Id: I8c4feb06e2129475ed6e31e3c4b4c4bc6fe1b27c
Reviewed-on: http://gerrit.chromium.org/gerrit/7422
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>