Fixing network_3GSuspendResume test
network_3GSuspendResume test has been aborting and failing for a couple of reasons
1) Device was not signed in when suspending/resuming the device which powered
   off the device
2) Device did not have the cellular service as a Favorite.

The fix should login to the device before performing any suspend/resume.
Secondly, it should check the remembered state of the cellular service.  If it is
not remembered yet, we connect to the service first before continuing with the test.

TEST=run_remote_test.sh network_3GSuspendResume on a newly imaged device.  Also test the same script from the login screen
BUG=chromium-os:19907,chromium-os:19126
STATUS=Fixed

Change-Id: I5728141bbbf4fdb5051479e659e2b401ca83d46b
Reviewed-on: http://gerrit.chromium.org/gerrit/7224
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Stanley Wong <stanleyw@chromium.org>
Tested-by: Stanley Wong <stanleyw@chromium.org>