autotest: Simplify desktopui_WindowManagerFocusNewWindows.
This makes the test just check that the _NET_ACTIVE_WINDOW
X property is set, which acts as a good proxy for "is the
window manager running and focusing windows?".

I'm doing this to avoid flakiness that I suspect is being
caused by a race between Chrome's first window getting
painted and the windows created by the testing code being
mapped.  The additional logic that this test was checking
(namely, that newly-mapped windows take the focus from
existing windows) is already covered in much more detail by
layout_manager2_test in the window manager repository.

BUG=chromium-os:19786
TEST=ran it successfully

Change-Id: I0bf089018776a7a7e856fbaf4f91066453c14ee5
Reviewed-on: http://gerrit.chromium.org/gerrit/7292
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>