Wait for login prompt to show up after logout.
When we log out from a session, we should wait for Chrome to start up
completely before we mark the test as 'finished'. This ensures that the next
test will safely be able to stop Chrome without triggering any races
in the startup code.

My earlier fix for this was insufficient, because it did not actually check for
the login prompt to show up, and just checked for whether the Chrome process
has shown up. This will need to be adjusted to deal with web login ui.

BUG=chromium-os:18919
TEST=Ran trybot run with VM tests.

Change-Id: I758d584b78c72effe47af97c1d6b07d8df0a8414
Reviewed-on: http://gerrit.chromium.org/gerrit/7468
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>