Further enhance coverage of security_RootCA.
The default openssl routine for the test "is cert X a trusted root
cert?" looks in /etc/ssl/certs for files with a basename created
with an sprintf format along the lines of either "%08lx.%d" or
"%08lx.r%d".  This changes the test for the SSL certs to look at the
files matching that pattern.

BUG=None
TEST=run the new test against build 0.15.1005.0

Change-Id: Ib4385e69a693f399552d6b3933bdc5160fa181ab
Reviewed-on: http://gerrit.chromium.org/gerrit/7490
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Jim Hebert <jimhebert@chromium.org>