Factory: Allow SD/USB tests to run if test-media is already inserted.
This change allows the test to run if the media is ejected and then
re-inserted. Previously the test would be stuck always looking for
the insertion of a storage device that was already present.

BUG=chromium-os:20569
TEST=Run test with media inserted, eject and reinsert. Observe test run.
Run test with media not inserted. Insert media and observe test run.

Change-Id: If03fe5488cc781e91c8051c6091e6eccf06d7179
Reviewed-on: http://gerrit.chromium.org/gerrit/7917
Reviewed-by: Nick Sanders <nsanders@chromium.org>
Commit-Ready: Dave Parker <dparker@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>