iwcap: unpack tarball into git
I need to clean up some of the iwcap build code, and rather than unpacking
the tarball, making the changes, repacking the tarball, and then committing
the tarball into the git tree, let's start with getting rid of the tarball.

So all this commit does is unpack the iwcap.tar.gz and check in the files
that result.

BUG=chromium-os:20636
TEST=`emerge-x86-generic autotest-deps` still works and iwcap binary in same place

Change-Id: I3742871db45ca2ca712547137994225c2f122efd
Reviewed-on: http://gerrit.chromium.org/gerrit/8073
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>